some small fixes

+++ src/main/java/com/example/task/request/TransactionRequest.java
@@ -1,10 +1,11 @@
 package com.example.task.request;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public record TransactionRequest (
         String name,
         LocalDateTime transactionDate,
-        Long amount,
+        BigDecimal amount,
         String transactionDetails
 ) {}
